+++ packages/spacecat-shared-data-access/test/unit/models/site/config.test.js
@@ -743,7 +743,7 @@ describe('Config Tests', () => {
         .to.throw().and.satisfy((error) => {
           expect(error.message).to.include('Configuration validation error');
           expect(error.cause.details[0].context.message)
-            .to.equal('"imports[0].destinations[0]" must be [default]. "imports[0].type" must be [organic-traffic]. "imports[0].type" must be [all-traffic]. "imports[0].type" must be [top-pages]. "imports[0].type" must be [cwv-daily]. "imports[0].type" must be [cwv-weekly]');
+            .to.equal('"imports[0].destinations[0]" must be [default]. "imports[0].type" must be [organic-keywords-nonbranded]. "imports[0].type" must be [organic-keywords-ai-overview]. "imports[0].type" must be [organic-keywords-feature-snippets]. "imports[0].type" must be [organic-keywords-questions]. "imports[0].type" must be [organic-traffic]. "imports[0].type" must be [all-traffic]. "imports[0].type" must be [top-pages]. "imports[0].type" must be [cwv-daily]. "imports[0].type" must be [cwv-weekly]');
           expect(error.cause.details[0].context.details)
             .to.eql([
               {
@@ -763,6 +763,74 @@ describe('Config Tests', () => {
                   value: 'invalid',
                   key: 0,
                 },
+              },
+              {
+                context: {
+                  key: 'type',
+                  label: 'imports[0].type',
+                  valids: [
+                    'organic-keywords-nonbranded',
+                  ],
+                  value: 'organic-keywords',
+                },
+                message: '"imports[0].type" must be [organic-keywords-nonbranded]',
+                path: [
+                  'imports',
+                  0,
+                  'type',
+                ],
+                type: 'any.only',
+              },
+              {
+                context: {
+                  key: 'type',
+                  label: 'imports[0].type',
+                  valids: [
+                    'organic-keywords-ai-overview',
+                  ],
+                  value: 'organic-keywords',
+                },
+                message: '"imports[0].type" must be [organic-keywords-ai-overview]',
+                path: [
+                  'imports',
+                  0,
+                  'type',
+                ],
+                type: 'any.only',
+              },
+              {
+                context: {
+                  key: 'type',
+                  label: 'imports[0].type',
+                  valids: [
+                    'organic-keywords-feature-snippets',
+                  ],
+                  value: 'organic-keywords',
+                },
+                message: '"imports[0].type" must be [organic-keywords-feature-snippets]',
+                path: [
+                  'imports',
+                  0,
+                  'type',
+                ],
+                type: 'any.only',
+              },
+              {
+                context: {
+                  key: 'type',
+                  label: 'imports[0].type',
+                  valids: [
+                    'organic-keywords-questions',
+                  ],
+                  value: 'organic-keywords',
+                },
+                message: '"imports[0].type" must be [organic-keywords-questions]',
+                path: [
+                  'imports',
+                  0,
+                  'type',
+                ],
+                type: 'any.only',
               },
               {
                 message: '"imports[0].type" must be [organic-traffic]',